move more items to new service w/ burnettk
